+++ src/test/groovy/ru/vyarus/dropwizard/guice/bundles/lookup/DefaultLookupTest.groovy
@@ -3,6 +3,8 @@ package ru.vyarus.dropwizard.guice.bundles.lookup
 import ru.vyarus.dropwizard.guice.AbstractTest
 import ru.vyarus.dropwizard.guice.bundle.DefaultBundleLookup
 import ru.vyarus.dropwizard.guice.bundle.lookup.PropertyBundleLookup
+import ru.vyarus.dropwizard.guice.module.installer.CoreInstallersBundle
+import ru.vyarus.dropwizard.guice.module.installer.bundle.GuiceyBundle
 import ru.vyarus.dropwizard.guice.module.jersey.debug.HK2DebugBundle
 
 /**
@@ -10,6 +12,19 @@ import ru.vyarus.dropwizard.guice.module.jersey.debug.HK2DebugBundle
  * @since 17.01.2016
  */
 class DefaultLookupTest extends AbstractTest {
+
+    String serviceFile = "build/classes/test/META-INF/services/${GuiceyBundle.class.name}"
+
+    void setup() {
+        new File(serviceFile).parentFile.mkdirs()
+    }
+
+    void cleanup() {
+        File file = new File(serviceFile)
+        if (file.exists()) {
+            file.delete()
+        }
+    }
 
     def "Check default lookup"() {
 
@@ -24,5 +39,14 @@ class DefaultLookupTest extends AbstractTest {
         then: "resolved"
         res.size() == 1
         res[0] instanceof HK2DebugBundle
+
+        when: "init from both loaders"
+        PropertyBundleLookup.enableBundles(HK2DebugBundle)
+        new File(serviceFile) << CoreInstallersBundle.class.name
+        res = new DefaultBundleLookup().lookup()
+        then: "resolved"
+        res.size() == 2
+        res[0] instanceof HK2DebugBundle
+        res[1] instanceof CoreInstallersBundle
     }
 }